+++ src/Greedy.java
@@ -1,7 +1,7 @@
 
 public class Greedy {
 	private int numJobs;
-	private int[][] jobs;
+	private double[][] jobs;
 	
 	public Greedy(ProblemInstance instance) {
 		numJobs = instance.getNumJobs();
@@ -13,8 +13,8 @@ public class Greedy {
 	// to use this as a subroutine for a search method
 	public Schedule getSchedule() {
 		int jobID = -1;
-		int jobLength = -1;
-		int jobDueTime = -1;
+		double jobLength = -1;
+		double jobDueTime = -1;
 		
 		for(int i = 0; i < numJobs; ++i){
 			if(jobDueTime == -1 || jobDueTime > jobs[i][1]){
@@ -31,8 +31,8 @@ public class Greedy {
 		if(s.getDepth() >= numJobs) return s;
 		
 		int jobID = -1;
-		int jobLength = -1;
-		int jobDueTime = -1;
+		double jobLength = -1;
+		double jobDueTime = -1;
 		
 		for(int i = 0; i < numJobs; ++i){
 			if(s.containsJob(i) == false && (jobDueTime == -1 || jobDueTime > jobs[i][1])){
